requirements.txt
requests
python-dotenv
openai
feedparser
